+++ awl/tests/test_admintools.py
@@ -1,12 +1,13 @@
 # awl.tests.test_admintools.py
-from django.test import TestCase
 from django.contrib.admin.utils import label_for_field
+from django.test import TestCase
+from django.utils.safestring import SafeData
 
 from screwdriver import parse_link
 
 from awl.waelsteng import AdminToolsMixin
-from awl.tests.models import Nested, Inner, Outer
-from awl.tests.admin import InnerAdmin, OuterAdmin
+from awl.tests.models import Author, Book, Chapter
+from awl.tests.admin import BookAdmin, ChapterAdmin
 
 # ============================================================================
 
@@ -15,42 +16,46 @@ class AdminToolsTest(TestCase, AdminToolsMixin):
         # setup the admin site and id
         self.initiate()
 
-        inner_admin = InnerAdmin(Inner, self.site)
-        outer_admin = OuterAdmin(Outer, self.site)
+        book_admin = BookAdmin(Book, self.site)
+        chapter_admin = ChapterAdmin(Chapter, self.site)
 
-        n1 = Nested.objects.create(name='n1')
-        i1 = Inner.objects.create(name='i1', nested=n1)
-        o1 = Outer.objects.create(name='o1', inner=i1)
+        tolstoy = Author.objects.create(name='Tolstoy')
+        war = Book.objects.create(name='War and Peace', author=tolstoy)
+        ch1 = Chapter.objects.create(name='Part I.I', book=war)
 
-        # check the basic __str__ named link from Inner to Nested
-        html = self.field_value(inner_admin, i1, 'show_nested')
+        # make sure we get a safe string
+        html = self.field_value(book_admin, war, 'show_author')
+        self.assertTrue(issubclass(html.__class__, SafeData))
+
+        # check the basic __str__ named link from Book to Author
+        html = self.field_value(book_admin, war, 'show_author')
         url, text = parse_link(html)
-        self.assertEqual('Nested(id=1 n1)', text)
-        self.assertEqual('/admin/tests/nested/?id__exact=1', url)
+        self.assertEqual('Author(id=1 Tolstoy)', text)
+        self.assertEqual('/admin/tests/author/?id__exact=1', url)
 
-        # check the template based name link of Inner from Outer
-        html = self.field_value(outer_admin, o1, 'show_inner')
+        # check the template based name link of Book from Chapter
+        html = self.field_value(chapter_admin, ch1, 'show_book')
         url, text = parse_link(html)
-        self.assertEqual('Inner.id=1', text)
-        self.assertEqual('/admin/tests/inner/?id__exact=1', url)
+        self.assertEqual('Book.id=1', text)
+        self.assertEqual('/admin/tests/book/?id__exact=1', url)
 
-        # check the double dereferenced Nested from Outer
-        html = self.field_value(outer_admin, o1, 'show_nested')
+        # check the double dereferenced Author from Chapter
+        html = self.field_value(chapter_admin, ch1, 'show_author')
         url, text = parse_link(html)
-        self.assertEqual('Nested(id=1 n1)', text)
-        self.assertEqual('/admin/tests/nested/?id__exact=1', url)
+        self.assertEqual('Author(id=1 Tolstoy)', text)
+        self.assertEqual('/admin/tests/author/?id__exact=1', url)
 
         # check the title got set correctly
-        label = label_for_field('show_inner', o1, outer_admin)
-        self.assertEqual(label, 'My Inner')
+        label = label_for_field('show_book', ch1, chapter_admin)
+        self.assertEqual(label, 'My Book')
 
         # check that empty values work properly
-        o2 = Outer.objects.create(name='o2')
-        result = self.field_value(outer_admin, o2, 'show_nested')
+        ch2 = Chapter.objects.create(name='Part I.II')
+        result = self.field_value(chapter_admin, ch2, 'show_author')
         self.assertEqual('', result)
 
-        i2 = Inner.objects.create(name='i2')
-        o2.inner = i2
-        o2.save()
-        result = self.field_value(outer_admin, o2, 'show_nested')
+        book = Book.objects.create(name='book')
+        ch2.book = book
+        ch2.save()
+        result = self.field_value(chapter_admin, ch2, 'show_author')
         self.assertEqual('', result)